requirements.txt
langchain-huggingface 
pinecone-client 
langchain
langchain_groq 
langchain-core 
langchain_community 
langchain_pinecone
